Release unit test cases, code formatting.

+++ src/main/java/org/capgemini/blue_harvest/accountservice/dao/CustomerDao.java
@@ -6,7 +6,8 @@ import java.util.Optional;
 import org.capgemini.blue_harvest.accountservice.entity.Customer;
 
 public interface CustomerDao {
-	public Optional<Customer> getCustomerById(long id);
+	public Optional<Customer> getCustomerById(int id);
+
 	public List<Customer> getAllCustomers();
 
 }
